requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# IBM Watsonx.ai SDK
ibm-watsonx-ai==0.1.0
ibm-cloud-sdk-core==3.18.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Configuration
pyyaml==6.0.1

# Logging and utilities
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: IBM ACP Gateway (if using official gateway)
# ibm-acp-gateway==1.0.0 